[559] Add Selection Dialog on Composite Processor Tool
First commit of the 'support for a selection dialog on diagram tools'.
When the user selects the Composite Processor Tool in the
Flow/Topography Diagram, a 'Selection Dialog' popup is displayed asking
for the selection of an object.
The objects are fake for this first commit. The list of objects will be
filled with real objects in the following commits associated to this
ticket.
The tool is only executed if the user select an object from the list and
clicks on the 'Finish' button, otherwise the tool is not executed.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/559
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>