WIP: [451] Fix support for label-edit enablement by direct typing
Sprotty does not support this mode, so when we trigger the
`EditLabelAction` in our keyboard listener, by default the first
character typed is lost.

The patch adds this `firstChar` property to the action, which is then
available in `SiriusWebWebSocketDiagramServer.handleEditLabelAction`.

Sprotty's normal behavior in `EditLabelUI.applyTextContents` is to
initialize the contents of the `<input/>` element with the SLabel's
`text`, and then to make the text selected.

The patch modifies the SLabel's `text` with the `fistChar` to get the
correct initial contents, and overrides `applyTextContents` to remove
the selection.

This has two drawbacks:
1. the initial text is no longer selected even for other direct edit
   methods (via F2 or the palette) where it still makes sense;
2. because we modify the SLabel's text, when aborting the direct edit
   that change is not reverted and the label wrongly displays the
   `firstChar` we set.

Both would be fixed if we could pass some additional information to
`EditLabelUI` to distinguish the different "modes", but it's not
clear (to me) how to do that given Sprotty's API/protocol.

One approach could be to use a subclass of `SLabel` with the
additional information we need, but it seems quite a big/impacting
change just for that simple case.

Note that the current patch only handles `firstChar`, but it
could/should be improved to support a more general `initialText` that
could be different from the label's text, which would enable
supporting something like Sirius Desktop's
`DirectEditLabel.inputLabelExpression`.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/451
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>